requirements.txt
streamlit
python-dotenv
requests
gtts

